// src/app/user/model/user.ts
export class User {
  id: number;
  password: string;
  last_login: string;
  is_superuser: boolean;
  full_name: string;
  first_name: string;
  last_name: string;
  email: string;
  is_staff: boolean;
  is_active: boolean;
  date_joined: string;
  username: string
  confirm_username: boolean;
  is_social: boolean;
  phone: string;
  publisher: boolean;
  name: string;
  cpf: string;
  rg: number;
  birth_date: string;
  lattes: string;
  about: string;
  groups: string[];
  user_permissions: string[]
}